Add ESC to call pauseMenu in game. 修复了 PauseMenu 中的全局变量使用错误。

+++ Classes/Actor/Controller.cpp
@@ -17,7 +17,7 @@ bool MoveController::init()
 	return true;
 }
 
-void MoveController::setSpeed(int speed)
+void MoveController::setSpeed(float speed)
 {
 	m_speed = speed;
 }
